Add files via upload
In this project, we aim to train machine learning algorithms on the MNIST dataset to create a machine capable of recognizing handwritten digits. We utilize three different algorithms for this purpose: KNN, CNN, and SVM.

KNN (k-Nearest Neighbors): This non-parametric classification algorithm classifies handwritten digits based on the labels of the nearest examples in the training dataset.

CNN (Convolutional Neural Network): Convolutional neural networks are particularly effective for image recognition. They use convolutional layers to extract important features from images of handwritten digits, followed by pooling layers and fully connected layers to perform the classification.

SVM (Support Vector Machine): Support vector machines find the optimal hyperplane that separates different classes in a high-dimensional space, making them effective for classification tasks such as recognizing handwritten digits.

The goal of this project is to compare the performance of these three algorithms on the MNIST dataset to determine which one is the most effective for recognizing handwritten digits.